verbose-equals-true/backend/requirements.txt
Django==2.2.10
psycopg2-binary
flake8==3.7.7
pytest 
pytest-cov
pytest-django
djangorestframework
django-filter
djangorestframework-jwt
gunicorn==19.8
celery
redis